Sounds added. Known issue: sometimes sounds play at the same time, like a gong and dying sound. Known issue: you don't hear sounds, besides the piece dying sound, from the enemy even if you can see him, like if he lands on a carrot.
Fix issue where multiple playing pieces could be on the same square. Game was checking only the first thing found on a destination square before performing a move. This was sometimes terrain, which pieces can be on top of, even if it would have found a playing piece if it searched all things on the destination square. 

Allow new units gained via carrots to appear on top of terrain.

Fix issue where a unit gained via a carrot could be born under a unit that just moved next to the home. Game was checking where to place the new unit before completing the move.

Maintain order of joinable games. The list visible on the start screen was changing order as you watched.

Remove games from the observable games list where you are a player, if you are logged in. Getting back to games you are playing was added to the joinable games list in an earlier revision. So this is no longer needed and confused testers.